% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codons}
\alias{codons}
\title{codons}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codons
}
\description{
eukaryotic codon translation table
}
\keyword{datasets}
