% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_aa.R
\name{count_aa}
\alias{count_aa}
\title{count_aa}
\usage{
count_aa(aa_seq)
}
\arguments{
\item{aa_seq}{Input is a string of amino acids in one letter code format.}
}
\value{
This function returns a visualization of the amino acid composition of the input sequence
}
\description{
count_aa
}
\examples{
aa_seq = "MGTCCEYKLDMPRQQCLKYWALKDRTYNV"
count_aa(aa_seq)
}
