% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_string.R
\name{create_string}
\alias{create_string}
\title{create_string}
\usage{
create_string(size)
}
\arguments{
\item{size}{Enter the length of the simulated DNA string you wish.}
}
\value{
string
}
\description{
create_string
}
\examples{
length <- 50
create_string(length)

}
