% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{dna_to_rna}
\usage{
dna_to_rna(seq)
}
\arguments{
\item{seq}{DNA sequence input}
}
\value{
reverse_seq DNA coding sequence input returned as RNA sequence
}
\description{
dna_to_rna
}
\examples{
sequence <- "GATCGATCGTAGCTAGCTGACTG"
dna_to_rna(sequence)
}
