% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_to_codons.R
\name{rna_to_codons}
\alias{rna_to_codons}
\title{RNA string to codons}
\usage{
rna_to_codons(rna_string, start = 1)
}
\arguments{
\item{rna_string}{Paste the name of the RNA string here.}

\item{start}{Choose start point for translation in RNA string. Default = 1.}
}
\value{
output. Codons corresponding to the RNA string.
}
\description{
RNA string to codons
}
\examples{
sequence <- "GUACUCUUCCAAGAUCUACUAGCUAGCUAUAGGG"
}
