% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(codon)
}
\arguments{
\item{codon}{vector of codons representing an RNA sequence}
}
\value{
returns collapsed amino acid sequence as a string
}
\description{
translate
}
\examples{
AAs <- c("GCG", "AUU", "GUU", "GAU", "UUU", "CGC", "CCU", "UCG", "AAC", "AUC", "GUU", "UAA")
translate(AAs)
}
