% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_string.R
\name{create_String}
\alias{create_String}
\title{Create a DNA string}
\usage{
create_String(size)
}
\arguments{
\item{size}{Enter the length of the simulated DNA string you wish.}
}
\value{
string
}
\description{
Create a DNA string
}
\examples{
length <- 50
create_String(length)

}
