% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(vector_of_codons)
}
\arguments{
\item{vector_of_codons}{vector of codons representing an RNA sequence}
}
\value{
returns collapsed amino acid sequence as a string
}
\description{
translate
}
\examples{
AAs <- c(M, A, G, G, T, Y, Y, P, I, K, L, M)
translate(AAs)
}
